requirements.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
djangorestframework==3.14.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
telegram==0.0.1
tzdata==2023.3
urllib3==2.0.3
